+++ src/lightning_habana/__init__.py
@@ -19,6 +19,7 @@ from lightning_utilities import compare_version
 from lightning_habana.__about__ import *  # noqa: F401, F403
 from lightning_habana.pytorch.accelerator import HPUAccelerator
 from lightning_habana.pytorch.datamodule.datamodule import HPUDataModule
+from lightning_habana.pytorch.plugins.deepspeed_precision import HPUDeepSpeedPrecisionPlugin
 from lightning_habana.pytorch.plugins.io_plugin import HPUCheckpointIO
 from lightning_habana.pytorch.plugins.precision import HPUPrecisionPlugin
 from lightning_habana.pytorch.profiler.profiler import HPUProfiler
@@ -53,4 +54,5 @@ __all__ = [
     "HPUDataModule",
     "HPUProfiler",
     "HPU_AVAILABLE",
+    "HPUDeepSpeedPrecisionPlugin",
 ]